# docs/killrvideo_openapi.yaml
openapi: 3.1.0
info:
  title: KillrVideo v2 - Monolith Backend
  version: 2.0.0-alpha-local
paths:
...
  /api/v1/flags:
    post:
      tags:
      - Flags
      summary: Flag content (video or comment)
      description: Create a new flag for the specified content.
      operationId: submit_flag_api_v1_flags_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlagCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlagResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/moderation/flags:
    get:
      tags:
      - Moderation Actions
      summary: List all flags (moderator inbox)
      operationId: list_all_flags_api_v1_moderation_flags_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/FlagStatusEnum'
          - type: 'null'
          description: Filter by flag status
          title: Status
        description: Filter by flag status
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 10
          title: Pagesize
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_FlagResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/moderation/flags/{flag_id_path}:
    get:
      tags:
      - Moderation Actions
      summary: Get details of a specific flag
      operationId: get_flag_details_api_v1_moderation_flags__flag_id_path__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: flag_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Flag Id Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlagResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/moderation/flags/{flag_id_path}/action:
    post:
      tags:
      - Moderation Actions
      summary: Take action on a specific flag
      operationId: act_on_flag_api_v1_moderation_flags__flag_id_path__action_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: flag_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Flag Id Path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlagUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlagResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/moderation/users:
    get:
      tags:
      - Moderation Actions
      summary: Search for users (moderator only)
      operationId: search_users_endpoint_api_v1_moderation_users_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search text
          title: Q
        description: Search text
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                title: Response Search Users Endpoint Api V1 Moderation Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/moderation/users/{user_id_path}/assign-moderator:
    post:
      tags:
      - Moderation Actions
      summary: Promote user to moderator
      operationId: assign_moderator_endpoint_api_v1_moderation_users__user_id_path__assign_moderator_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/moderation/users/{user_id_path}/revoke-moderator:
    post:
      tags:
      - Moderation Actions
      summary: Demote user from moderator
      operationId: revoke_moderator_endpoint_api_v1_moderation_users__user_id_path__revoke_moderator_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/moderation/videos/{video_id_path}/restore:
    post:
      tags:
      - Moderation Actions
      summary: Restore a soft-deleted video (stub)
      operationId: restore_video_endpoint_api_v1_moderation_videos__video_id_path__restore_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: video_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Video Id Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentRestoreResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/moderation/comments/{comment_id_path}/restore:
    post:
      tags:
      - Moderation Actions
      summary: Restore a soft-deleted comment (stub)
      operationId: restore_comment_endpoint_api_v1_moderation_comments__comment_id_path__restore_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: comment_id_path
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Comment Id Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentRestoreResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ContentRestoreResponse:
      properties:
        content_id:
          type: string
          format: uuid
          title: Content Id
        content_type:
          type: string
          title: Content Type
        status_message:
          type: string
          title: Status Message
      type: object
      required:
      - content_id
      - content_type
      - status_message
      title: ContentRestoreResponse
    ContentTypeEnum:
      type: string
      enum:
      - video
      - comment
      title: ContentTypeEnum
      description: Possible kinds of content that can be flagged.
...
    FlagCreateRequest:
      properties:
        contentType:
          $ref: '#/components/schemas/ContentTypeEnum'
        contentId:
          type: string
          format: uuid
          title: Contentid
        reasonCode:
          $ref: '#/components/schemas/FlagReasonCodeEnum'
        reasonText:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Reasontext
          description: Optional free-form context supplied by the reporter.
      type: object
      required:
      - contentType
      - contentId
      - reasonCode
      title: FlagCreateRequest
      description: Payload viewers submit when flagging content.
    FlagReasonCodeEnum:
      type: string
      enum:
      - spam
      - inappropriate
      - harassment
      - copyright
      - other
      title: FlagReasonCodeEnum
      description: Standardized reasons a user can provide when flagging content.
    FlagResponse:
      properties:
        contentType:
          $ref: '#/components/schemas/ContentTypeEnum'
        contentId:
          type: string
          format: uuid
          title: Contentid
        reasonCode:
          $ref: '#/components/schemas/FlagReasonCodeEnum'
        reasonText:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Reasontext
          description: Optional free-form context supplied by the reporter.
        flagId:
          type: string
          format: uuid
          title: Flagid
        userId:
          type: string
          format: uuid
          title: Userid
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
        status:
          $ref: '#/components/schemas/FlagStatusEnum'
          default: open
        moderatorId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Moderatorid
        moderatorNotes:
          anyOf:
          - type: string
          - type: 'null'
          title: Moderatornotes
        resolvedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Resolvedat
      type: object
      required:
      - contentType
      - contentId
      - reasonCode
      - userId
      - createdAt
      - updatedAt
      title: FlagResponse
      description: Alias used in API responses.
    FlagStatusEnum:
      type: string
      enum:
      - open
      - under_review
      - approved
      - rejected
      title: FlagStatusEnum
      description: Lifecycle states for the moderation flag itself.
    FlagUpdateRequest:
      properties:
        status:
          $ref: '#/components/schemas/FlagStatusEnum'
        moderatorNotes:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Moderatornotes
      type: object
      required:
      - status
      title: FlagUpdateRequest
      description: Payload moderators send when changing a flag's status.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedResponse_FlagResponse_:
      properties:
        data:
          items:
            $ref: '#/components/schemas/FlagResponse'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - data
      - pagination
      title: PaginatedResponse[FlagResponse]
...
    Pagination:
      properties:
        currentPage:
          type: integer
          title: Currentpage
        pageSize:
          type: integer
          title: Pagesize
        totalItems:
          type: integer
          title: Totalitems
        totalPages:
          type: integer
          title: Totalpages
      type: object
      required:
      - currentPage
      - pageSize
      - totalItems
      - totalPages
      title: Pagination
...
    User:
      properties:
        firstName:
          type: string
          maxLength: 50
          minLength: 1
          title: Firstname
        lastName:
          type: string
          maxLength: 50
          minLength: 1
          title: Lastname
        email:
          type: string
          format: email
          title: Email
        userId:
          type: string
          format: uuid
          title: Userid
        createdDate:
          type: string
          format: date-time
          title: Createddate
        accountStatus:
          type: string
          title: Accountstatus
          default: active
        lastLoginDate:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Lastlogindate
        roles:
          items:
            type: string
          type: array
          title: Roles
          default: []
      type: object
      required:
      - firstName
      - lastName
      - email
      - userId
      title: User
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/users/login